Refactor : 도메인 설계 변경
