+++ src/main/java/umc/spring/web/controller/StoreRestController.java
@@ -74,5 +74,4 @@ public class StoreRestController {
     }
 
 
-
 }
